requirements/base.txt
Django==2.2.1
django-environ==0.4.5
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
